refactor: move orca.tsx logic into App.tsx
